sensor-daemon/requirements.txt
python-daemon
hiredis
redis